+++ src/NServiceBus.Msmq.AcceptanceTests/NServiceBus.Msmq.AcceptanceTests.csproj
@@ -69,11 +69,11 @@
     <Compile Include="When_distributing_a_command.cs" />
     <Compile Include="When_overriding_local_address.cs" />
     <Compile Include="When_not_using_instance_mapping_file.cs" />
-    <Compile Include="When_subscription_exists_for_non_tx_endpoint.cs" />
-    <Compile Include="When_using_subscription_store_on_non_tx_endpoint.cs" />
-    <Compile Include="When_subscription_exists_for_tx_endpoint.cs" />
-    <Compile Include="When_using_subscription_store_on_tx_endpoint.cs" />
-    <Compile Include="When_publishing.cs" />
+    <Compile Include="SubscriptionStorage\When_subscription_exists_for_non_tx_endpoint.cs" />
+    <Compile Include="SubscriptionStorage\When_using_subscription_store_on_non_tx_endpoint.cs" />
+    <Compile Include="SubscriptionStorage\When_subscription_exists_for_tx_endpoint.cs" />
+    <Compile Include="SubscriptionStorage\When_using_subscription_store_on_tx_endpoint.cs" />
+    <Compile Include="SubscriptionStorage\When_publishing.cs" />
     <Compile Include="When_publishing_with_authorizer.cs" />
     <Compile Include="When_receiving_control_message_with_body.cs" />
     <Compile Include="When_receiving_with_dtc_disabled.cs" />
