+++ src/NServiceBus.Msmq.AcceptanceTests/SubscriptionStorage/When_subscription_exists_for_tx_endpoint.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests
+﻿namespace NServiceBus.AcceptanceTests.SubscriptionStorage
 {
     using System;
     using System.Messaging;
