+++ src/NServiceBus.Msmq.AcceptanceTests/SubscriptionStorage/When_using_subscription_store_on_non_tx_endpoint.cs
@@ -1,20 +1,19 @@
-﻿namespace NServiceBus.AcceptanceTests
+﻿namespace NServiceBus.AcceptanceTests.SubscriptionStorage
 {
     using System;
     using System.Messaging;
     using System.Threading.Tasks;
     using AcceptanceTesting;
-    using AcceptanceTesting.Customization;
     using Config;
     using Config.ConfigurationSource;
     using Features;
     using NUnit.Framework;
     using Persistence.Legacy;
 
-    public class When_subscription_exists_for_non_tx_endpoint : NServiceBusAcceptanceTest
+    public class When_using_subscription_store_on_non_tx_endpoint : NServiceBusAcceptanceTest
     {
         [Test]
-        public async Task Should_use_subscription()
+        public async Task Should_persist_subscriptions()
         {
             var queuePath = $".\\private$\\{StorageQueueName}";
 
@@ -23,22 +22,11 @@
                 MessageQueue.Delete(queuePath);
             }
 
-            MessageQueue.Create(queuePath, false);
-
-            using (var queue = new MessageQueue(queuePath))
-            {
-                queue.Send(new Message
-                {
-                    Label = Conventions.EndpointNamingConvention(typeof(Subscriber)),
-                    Body = typeof(MyEvent).AssemblyQualifiedName
-                }, MessageQueueTransactionType.None);
-            }
-
             var ctx = await Scenario.Define<Context>()
                 .WithEndpoint<Publisher>(b =>
-                            b.When(c => c.EndpointsStarted, (session, c) => session.Publish(new MyEvent()))
+                            b.When(c => c.Subscribed, (session, c) => session.Publish(new MyEvent()))
                 )
-                .WithEndpoint<Subscriber>()
+                .WithEndpoint<Subscriber>(b => b.When(session => session.Subscribe<MyEvent>()))
                 .Done(c => c.GotTheEvent)
                 .Run(TimeSpan.FromSeconds(10));
 
@@ -50,11 +38,12 @@
             }
         }
 
-        static string StorageQueueName = "msmq.acpt.nontxexistingsubscriptions";
+        static string StorageQueueName = "msmq.acpt.nontxsubscriptions";
 
         public class Context : ScenarioContext
         {
             public bool GotTheEvent { get; set; }
+            public bool Subscribed { get; set; }
         }
 
         public class Publisher : EndpointConfigurationBuilder
@@ -63,6 +52,7 @@
             {
                 EndpointSetup<DefaultPublisher>(b =>
                 {
+                    b.OnEndpointSubscribed<Context>((s, context) => { context.Subscribed = true; });
                     b.DisableFeature<AutoSubscribe>();
                     b.UsePersistence<MsmqPersistence>();
                     b.UseTransport<MsmqTransport>()
