+++ src/NServiceBus.Msmq.AcceptanceTests/SubscriptionStorage/When_using_subscription_store_on_tx_endpoint.cs
@@ -1,43 +1,44 @@
-﻿namespace NServiceBus.AcceptanceTests
+﻿namespace NServiceBus.AcceptanceTests.SubscriptionStorage
 {
     using System;
+    using System.Messaging;
     using System.Threading.Tasks;
     using AcceptanceTesting;
+    using Config;
+    using Config.ConfigurationSource;
     using Features;
-    using Persistence.Legacy;
     using NUnit.Framework;
+    using Persistence.Legacy;
 
-    public class When_using_subscription_store : NServiceBusAcceptanceTest
+    public class When_using_subscription_store_on_tx_endpoint : NServiceBusAcceptanceTest
     {
         [Test]
-        public async Task Should_be_delivered_to_all_subscribers()
+        public async Task Should_persist_subscriptions()
         {
+            var queuePath = $".\\private$\\{StorageQueueName}";
+
+            if (MessageQueue.Exists(queuePath))
+            {
+                MessageQueue.Delete(queuePath);
+            }
+
             var ctx = await Scenario.Define<Context>()
                 .WithEndpoint<Publisher>(b =>
-                    b.When(c => c.Subscribed, (session, c) =>
-                    {
-                        c.AddTrace("Both subscribers is subscribed, going to publish MyEvent");
-                        return session.Publish(new MyEvent());
-                    })
+                            b.When(c => c.Subscribed, (session, c) => session.Publish(new MyEvent()))
                 )
-                .WithEndpoint<Subscriber>(b => b.When(async (session, context) =>
-                {
-                    await session.Subscribe<MyEvent>();
-                    if (context.HasNativePubSubSupport)
-                    {
-                        context.Subscribed = true;
-                        context.AddTrace("Subscriber1 is now subscribed (at least we have asked the broker to be subscribed)");
-                    }
-                    else
-                    {
-                        context.AddTrace("Subscriber1 has now asked to be subscribed to MyEvent");
-                    }
-                }))
+                .WithEndpoint<Subscriber>(b => b.When(session => session.Subscribe<MyEvent>()))
                 .Done(c => c.GotTheEvent)
                 .Run(TimeSpan.FromSeconds(10));
 
             Assert.IsTrue(ctx.GotTheEvent);
+
+            using (var queue = new MessageQueue(queuePath))
+            {
+                CollectionAssert.IsNotEmpty(queue.GetAllMessages());
+            }
         }
+
+        static string StorageQueueName = "msmq.acpt.txsubscriptions";
 
         public class Context : ScenarioContext
         {
@@ -51,14 +52,21 @@
             {
                 EndpointSetup<DefaultPublisher>(b =>
                 {
-                    b.OnEndpointSubscribed<Context>((s, context) =>
-                    {
-                        context.Subscribed = true;
-                        context.AddTrace("Subscriber1 is now subscribed");
-                    });
+                    b.OnEndpointSubscribed<Context>((s, context) => { context.Subscribed = true; });
                     b.DisableFeature<AutoSubscribe>();
                     b.UsePersistence<MsmqPersistence>();
                 });
+            }
+
+            class QueueNameOverride : IProvideConfiguration<MsmqSubscriptionStorageConfig>
+            {
+                public MsmqSubscriptionStorageConfig GetConfiguration()
+                {
+                    return new MsmqSubscriptionStorageConfig
+                    {
+                        Queue = StorageQueueName
+                    };
+                }
             }
         }
 
@@ -82,7 +90,6 @@
             }
         }
 
-        [Serializable]
         public class MyEvent : IEvent
         {
         }
